requirements.txt
python-telegram-bot==20.3
deep-translator
nest_asyncio
aiogram==3.26.0
